Moved some source-handling stuff into `Preprocessor`
